refactor: OrderTable -> Order 패키지 의존성 인터페이스로 분리
